Fixed free (channel) selection in the signal selection panel.
There was a "Double cannot be cast to Float" exception in the
MP Decomposition configuration dialog.

+++ svarog/src/main/java/org/signalml/app/view/components/SignalSelectionPanel.java
@@ -439,8 +439,8 @@ public class SignalSelectionPanel extends JPanel {
 
 		} else if (signalSelectionTypePanel.getChannelRadio().isSelected()) {
 
-			float startTime = (Float) channelSignalSelectionPanel.getStartTimeSpinner().getValue();
-			float length = (Float) channelSignalSelectionPanel.getLengthSpinner().getValue();
+			double startTime = (Double) channelSignalSelectionPanel.getStartTimeSpinner().getValue();
+			double length = (Double) channelSignalSelectionPanel.getLengthSpinner().getValue();
 
 			selection = new SignalSelection(SignalSelectionType.CHANNEL);
 			selection.setPosition(startTime);
